REFACTOR: navbar 렌더링 계정확인 로직 분리

+++ src/views/account/order-view.js
@@ -1,47 +1,43 @@
 import * as Api from '/api.js';
+import checkAccount from '../check-account.js';
+import drawNavbar from '../navbar/index.js';
 
+const ref = {
+    cartCount: document.getElementById('cart-count'),
+};
 
 const userId = sessionStorage.getItem('userId');
 const productsContainer = document.querySelector('#productsContainer');
 
-
 async function fetchOrderInfo(userId) {
-
     // 테스트용: 6291d6e14cc1920b02fb4ce1
     const orderList = await Api.get('/api/auth', `${userId}/orders`);
 
-
-    const extractOrderInfo = orderList.map(orderInfo => {
-        return [
-            orderInfo.createdAt,
-            orderInfo._id,
-            orderInfo.paymentStatus,
-            [...orderInfo.productList]
-        ];
+    const extractOrderInfo = orderList.map((orderInfo) => {
+        return [orderInfo.createdAt, orderInfo._id, orderInfo.paymentStatus, [...orderInfo.productList]];
     });
 
-
     extractOrderInfo.forEach(async (orderItemInfo) => {
-        const [orderDate, orderId, orderState, itemList] = [orderItemInfo[0],
-        orderItemInfo[1],
-        orderItemInfo[2],
-        orderItemInfo[3]
+        const [orderDate, orderId, orderState, itemList] = [
+            orderItemInfo[0],
+            orderItemInfo[1],
+            orderItemInfo[2],
+            orderItemInfo[3],
         ];
         const orderDateForm = orderDate.slice(0, 10);
 
         // 주문 상태에 따라 구분
-        const orderProgress = orderState === 'Ok' ?
-                            '상품 준비중' : '결제완료';   
+        const orderProgress = orderState === 'Ok' ? '상품 준비중' : '결제완료';
 
         // 주문 내역의 아이템 이름, 수량
         const productString = [];
 
         itemList.forEach((obj) => {
             if (obj.name !== undefined) {
-                const info = {name: obj.name, quantity: obj.quantity};
+                const info = { name: obj.name, quantity: obj.quantity };
                 productString.push(info);
             }
-        })
+        });
 
         if (productString.length !== 0) {
             productString.forEach((productInfo) => {
@@ -56,15 +52,12 @@ async function fetchOrderInfo(userId) {
                             </div>`;
 
                 productsContainer.insertAdjacentHTML('beforeend', orderInfo);
-            })
+            });
         }
-
     });
 }
 
-
 await fetchOrderInfo(userId);
-
 
 const removeBtn = document.querySelectorAll('.button');
 
@@ -73,24 +66,16 @@ removeBtn.forEach((button) => {
         const orderId = e.target.id.slice(13);
         deleteOrder(button, orderId);
     });
-})
-
-
-
+});
 
 async function deleteOrder(button, orderId) {
     const entireRow = button.parentNode.parentNode;
-    const check = confirm(
-        '한 번 삭제한 주문은 복구가 불가능합니다. 그래도 삭제하시겠습니까?',
-    );
+    const check = confirm('한 번 삭제한 주문은 복구가 불가능합니다. 그래도 삭제하시겠습니까?');
 
     if (check) {
         productsContainer.removeChild(entireRow);
 
-        const result = await Api.delete(
-            '',
-            `api/auth/${userId}/orders/${orderId}`
-        );
+        const result = await Api.delete('', `api/auth/${userId}/orders/${orderId}`);
 
         if (result) {
             alert('삭제가 완료되었습니다.');
@@ -98,8 +83,26 @@ async function deleteOrder(button, orderId) {
     }
 }
 
+const drawCartCount = () => {
+    const cart = JSON.parse(localStorage.getItem('cart'));
 
+    if (!cart) {
+        ref.cartCount.innerText = 0;
+    } else {
+        ref.cartCount.innerText = Object.keys(cart).length;
+    }
+};
 
+const App = async () => {
+    const { isLogined, isAdmin } = await checkAccount();
 
+    if (!isLogined) {
+        alert('로그인이 필요한 페이지 입니다...');
+        window.location.href = '/';
+    }
 
+    drawNavbar('account', isLogined, isAdmin);
+    drawCartCount();
+};
 
+App();
